+++ Native/HookCommon/HookIpcProtocol.h
@@ -14,9 +14,6 @@ namespace ht::hook::ipc
     constexpr std::uint32_t kConfigHeaderMagic = 0x48435446; // "HCTF"
     constexpr std::uint32_t kConfigHeaderVersion = 1;
 
-    constexpr std::uint32_t kOverlayCmdMagic = 0x48434D44; // "HCMD"
-    constexpr std::uint32_t kOverlayCmdVersion = 1;
-
     // Overlay v2 (ImGui translation overlay): text blocks + UTF-8 blob.
     constexpr std::uint32_t kOverlayV2Magic = 0x32564F48; // "HOV2"
     constexpr std::uint32_t kOverlayV2Version = 2;
@@ -70,29 +67,6 @@ namespace ht::hook::ipc
         std::uint32_t reserved0;
         std::uint32_t reserved1;
         std::uint64_t updatedQpc;
-    };
-
-    struct OverlayCommandHeader
-    {
-        std::uint32_t magic;
-        std::uint32_t version;
-        std::uint32_t api;
-        std::uint32_t targetPid;
-        std::uint32_t commandCount;
-        std::uint32_t payloadBytes;
-        std::uint32_t reserved0;
-        std::uint32_t reserved1;
-        std::uint64_t updatedQpc;
-    };
-
-    struct OverlayRectCommand
-    {
-        float x;
-        float y;
-        float w;
-        float h;
-        std::uint32_t argb;
-        std::uint32_t thickness;
     };
 
     struct OverlayV2Header
@@ -165,15 +139,6 @@ namespace ht::hook::ipc
     inline std::wstring BuildConfigMappingName(DWORD pid, GraphicsApi api)
     {
         std::wstring name = L"Local\\HT_HOOK_CFG_";
-        name += std::to_wstring(static_cast<std::uint32_t>(api));
-        name += L"_";
-        name += std::to_wstring(pid);
-        return name;
-    }
-
-    inline std::wstring BuildOverlayCommandMappingName(DWORD pid, GraphicsApi api)
-    {
-        std::wstring name = L"Local\\HT_HOOK_CMD_";
         name += std::to_wstring(static_cast<std::uint32_t>(api));
         name += L"_";
         name += std::to_wstring(pid);